requirements.txt
streamlit
selenium
webdriver-manager
pillow
